[AGILE-236] Move shared WpSelect methods to concern
