+++ modules/backlogs/lib/open_project/backlogs/work_package_select_concern.rb
@@ -1,0 +1,62 @@
+# frozen_string_literal: true
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+module OpenProject::Backlogs
+  module WorkPackageSelectConcern
+    extend ActiveSupport::Concern
+
+    class_methods do
+      def instances(context = nil)
+        return [] if context && !context.backlogs_enabled?
+        return [] unless user_allowed_to_select?(context)
+
+        [new(context)]
+      end
+
+      def user_allowed_to_select?(context)
+        if context
+          User.current.allowed_in_project?(:view_sprints, context)
+        else
+          User.current.allowed_in_any_project?(:view_sprints)
+        end
+      end
+    end
+
+    private
+
+    def projects_with_view_sprints
+      if project
+        Project.where(id: project)
+      else
+        Project.allowed_to(User.current, :view_sprints)
+      end
+    end
+  end
+end
